phase1/synthesizer/requirements.txt
kafka-python==2.0.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
requests==2.31.0
fastapi==0.110.0
uvicorn==0.29.0